+++ csrc/runtime/hello.c
@@ -1,7 +1,0 @@
-#include <stdio.h>
-
-int main()
-{
-    printf("Hello from Soplang in C!\n");
-    return 0;
-}