+++ psrc/tests/__init__.py
@@ -1,1 +1,0 @@
-# Make tests a proper Python package